Frontend now uses the new java menu system
